+++ tests/test_play.py
@@ -1,17 +1,14 @@
+from .helpers import check_for_output
 import time
 from unittest import TestCase, skipUnless, mock
 from pya import *
 import numpy as np
-import pyaudio
 import warnings
+import pytest
+
 
 # check if we have an output device
-has_output = False
-try:
-    pyaudio.PyAudio().get_default_output_device_info()
-    has_output = True
-except OSError:
-    pass
+has_output = check_for_output()
 
 
 class MockAudio(mock.MagicMock):
@@ -35,18 +32,14 @@ class TestPlayBase(TestCase):
         self.sig2ch = np.repeat(self.sig, 2).reshape((44100, 2))
         self.astereo = Asig(self.sig2ch, sr=44100, label="stereo", cn=['l', 'r'])
 
-    def test_play(self):
-        s = Aserver(backend=self.backend)
-        s.boot()
-        self.asine.play(server=s)
+    @pytest.mark.xfail(reason="Test may get affect with PortAudio bug or potential unsuitable audio device.")
+    def test_play_and_stop(self):
+        ser = Aserver(backend=self.backend)
+        ser.boot()
+        self.asine.play(server=ser)
         time.sleep(2)
-
-    def test_stop(self):
-        s = Aserver(backend=self.backend)
-        s.boot()
-        self.asine.play(server=s)
-        s.stop()
-        s.quit()
+        ser.stop()
+        ser.quit()
 
     def test_gain(self):
         result = (self.asine * 0.2).sig
@@ -69,29 +62,29 @@ class TestPlay(TestPlayBase):
     __test__ = True
 
 
-class MockAudioTest(TestCase):
+# class MockAudioTest(TestCase):
 
-    # @skipUnless(has_output, "PyAudio found no output device.")
-    def test_play(self):
-        # Shift a mono signal to chan 4 should result in a 4 channels signals
-        mock_audio = MockAudio()
-        with mock.patch('pyaudio.PyAudio', return_value=mock_audio):
-            s = Aserver()
-            s.boot()
-            assert mock_audio.open.called
-            # since default AServer channel output is stereo we expect open to be called with
-            # channels=2
-            self.assertEqual(mock_audio.open.call_args_list[0][1]["channels"], 2)
-            d1 = np.linspace(0, 1, 44100)
-            d2 = np.linspace(0, 1, 44100)
-            asig = Asig(d1)
-            s.play(asig)
-            self.assertTrue(np.allclose(s.srv_asigs[0].sig, d2.reshape(44100, 1)))
+#     # @skipUnless(has_output, "PyAudio found no output device.")
+#     def test_play(self):
+#         # Shift a mono signal to chan 4 should result in a 4 channels signals
+#         mock_audio = MockAudio()
+#         with mock.patch('pyaudio.PyAudio', return_value=mock_audio):
+#             s = Aserver()
+#             s.boot()
+#             assert mock_audio.open.called
+#             # since default AServer channel output is stereo we expect open to be called with
+#             # channels=2
+#             self.assertEqual(mock_audio.open.call_args_list[0][1]["channels"], 2)
+#             d1 = np.linspace(0, 1, 44100)
+#             d2 = np.linspace(0, 1, 44100)
+#             asig = Asig(d1)
+#             s.play(asig)
+#             self.assertTrue(np.allclose(s.srv_asigs[0].sig, d2.reshape(44100, 1)))
 
-        with mock.patch('pyaudio.PyAudio', return_value=mock_audio):
-            with warnings.catch_warnings(record=True):
-                s = Aserver(channels=6)
-                s.boot()
-                assert mock_audio.open.call_count == 2
-                self.assertEqual(mock_audio.open.call_args_list[1][1]["channels"], 4)
+#         with mock.patch('pyaudio.PyAudio', return_value=mock_audio):
+#             with warnings.catch_warnings(record=True):
+#                 s = Aserver(channels=6)
+#                 s.boot()
+#                 assert mock_audio.open.call_count == 2
+#                 self.assertEqual(mock_audio.open.call_args_list[1][1]["channels"], 6)